Cut main.go in pieces, did the same with documentation
