+++ build.sh
@@ -1,4 +1,15 @@
 #!/bin/bash
+
+echo "----------------------------------------------------------------------"
+echo " go get -u"
+echo "----------------------------------------------------------------------"
+go get -u ./...
+
+echo "----------------------------------------------------------------------"
+echo " go mod tidy"
+echo "----------------------------------------------------------------------"
+go mod tidy
+
 rm bin/pgSimload*
 echo "----------------------------------------------------------------------"
 echo "Building binaries"
